add gauge to indexhtml

+++ ProjectBravo/IFluentGit.cs
@@ -29,6 +29,6 @@ namespace ProjectBravo
         Task<string> ThenReturnAuthorString();
         Task<List<FrequencyDTO>> ThenReturnFrequencyDTOList();
 
-        Task<int> ThenReturnAverageCommitsPerAuthorAsync();
+        Task<float> ThenReturnAverageCommitsPerAuthorAsync();
     }
 }
